feat(home): recent-page actions menu + active folder in directory sidebar
- Extract WikiItemActionsProvider + WikiItemMenu so the directory sidebar
  and the recent-pages grid share one contextual menu and one set of
  overlays (Share / Rename / Move / Copy Link / Launch Agent / Delete).
- Recent-page card ⋯ now opens that menu (was inert); wrapped in a span
  trigger since OPAL Button doesn't forward refs into Popover.Trigger.
- Sidebar ⋯ swaps in over the leading file/folder icon on hover.
- Active folder: clicking a folder row marks it active (highlighted);
  new pages/folders create inside it; reset when it (or an ancestor) is
  deleted. State lives in the provider so deletes can clear it.
