+++ frontend/src/components/wiki/WikiTree.module.css
@@ -66,17 +66,22 @@
   box-sizing: border-box;
 }
 
-/* Row wrapper — holds the SidebarTab plus the hover-revealed "⋯" menu, which
-   overlays the row's right edge (panel-coloured so it sits over the label). */
+/* Row wrapper — SidebarTab with the "⋯" menu overlaying its leading icon. */
 .rowWrap {
   position: relative;
   width: 100%;
 }
+.tab {
+  width: 100%;
+}
+/* "⋯" overlay — pinned over the leading icon slot (container p-2 = 8px + icon
+   box), hidden until the row is hovered or its menu is open. */
 .rowMenu {
   position: absolute;
-  right: 4px;
+  left: 6px;
   top: 50%;
   transform: translateY(-50%);
+  z-index: 1;
   opacity: 0;
   transition: opacity 0.12s ease;
 }
@@ -85,57 +90,31 @@
 .rowMenu[data-open] {
   opacity: 1;
 }
-/* "⋯" More button (657:29879): 24px, white surface (tint/00 → bg-page),
-   rounded-08 (md). Sits over the row's right edge. */
+/* Hide the file/folder icon while the "⋯" is showing, so it reads as a swap.
+   Scoped to .tab so the menu's own icon (in .rowMenu) is unaffected. */
+.rowWrap:hover .tab :global(svg),
+.rowWrap:has(.rowMenu[data-open]) .tab :global(svg) {
+  visibility: hidden;
+}
+/* "⋯" trigger — sized to sit in the icon slot, transparent so it reads as the
+   swapped-in icon (the real icon is hidden underneath while hovered). */
 .moreBtn {
   display: inline-flex;
   align-items: center;
   justify-content: center;
-  width: 24px;
-  height: 24px;
+  width: 20px;
+  height: 20px;
   padding: 0;
   border: none;
   border-radius: var(--border-radius-08);
-  background: var(--background-tint-00);
+  background: transparent;
   color: var(--text-03);
   cursor: pointer;
 }
 .moreBtn:hover {
-  background: var(--background-tint-03);
-  color: var(--text-04);
-}
-
-/* Contextual menu (657:41564): fixed 160px width, line items in a column. */
-.menu {
-  display: flex;
-  flex-direction: column;
-  gap: 2px;
-  width: 160px;
-  box-sizing: border-box;
-}
-/* Delete item — danger red (#dc2626 → --status-text-error-05) on icon + label.
-   !important to override OPAL LineItemButton's own text/icon color; scoped to
-   the single wrapped Delete item. */
-.danger,
-.danger :global(*) {
-  color: var(--status-text-error-05) !important;
+  color: var(--text-05);
 }
 
 .empty {
   padding: 8px 12px;
-}
-
-/* Transient toast (e.g. "Link copied") — bottom-centre, auto-dismisses. */
-.toast {
-  position: fixed;
-  left: 50%;
-  bottom: 32px;
-  transform: translateX(-50%);
-  z-index: var(--z-toast);
-  padding: 8px 14px;
-  border-radius: var(--border-radius-08);
-  background: var(--background-tint-04);
-  color: var(--text-05);
-  font-size: 13px;
-  box-shadow: var(--shadow-popover);
 }
